+++ Packages/UGF.Module.Assets/Runtime/AssetGroupAsset.cs
@@ -1,29 +1,18 @@
-﻿using UGF.EditorTools.Runtime.IMGUI.Attributes;
+﻿using System;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace UGF.Module.Assets.Runtime
 {
-    public abstract class AssetGroupAsset : AssetGroupAssetBase
+    public abstract class AssetGroupAsset : ScriptableObject
     {
-        [AssetGuid(typeof(AssetLoaderAssetBase))]
-        [SerializeField] private string m_loader;
-
-        public string Loader { get { return m_loader; } set { m_loader = value; } }
-
-        protected override IAssetGroup OnBuild()
+        public void GetAssets(IDictionary<string, IAssetInfo> assets)
         {
-            IAssetGroup group = OnBuildGroup();
+            if (assets == null) throw new ArgumentNullException(nameof(assets));
 
-            OnPopulateGroup(group);
-
-            return group;
+            OnGetAssets(assets);
         }
 
-        protected virtual IAssetGroup OnBuildGroup()
-        {
-            return new AssetGroup(m_loader);
-        }
-
-        protected abstract void OnPopulateGroup(IAssetGroup group);
+        protected abstract void OnGetAssets(IDictionary<string, IAssetInfo> assets);
     }
 }