+++ Packages/UGF.Module.Assets/Runtime/AssetInfo.cs
@@ -4,12 +4,15 @@ namespace UGF.Module.Assets.Runtime
 {
     public class AssetInfo : IAssetInfo
     {
+        public string LoaderId { get; }
         public string Address { get; }
 
-        public AssetInfo(string address)
+        public AssetInfo(string loaderId, string address)
         {
+            if (string.IsNullOrEmpty(loaderId)) throw new ArgumentException("Value cannot be null or empty.", nameof(loaderId));
             if (string.IsNullOrEmpty(address)) throw new ArgumentException("Value cannot be null or empty.", nameof(address));
 
+            LoaderId = loaderId;
             Address = address;
         }
     }