+++ Packages/UGF.Module.Assets/Runtime/IAssetModuleDescription.cs
@@ -3,10 +3,10 @@ using UGF.Application.Runtime;
 
 namespace UGF.Module.Assets.Runtime
 {
-    public interface IAssetsModuleDescription : IApplicationModuleDescription
+    public interface IAssetModuleDescription : IApplicationModuleDescription
     {
         IReadOnlyDictionary<string, IAssetLoader> Loaders { get; }
-        IReadOnlyDictionary<string, IAssetGroup> Groups { get; }
+        IReadOnlyDictionary<string, IAssetInfo> Assets { get; }
         IReadOnlyList<string> PreloadAssets { get; }
         IReadOnlyList<string> PreloadAssetsAsync { get; }
         bool UnloadTrackedAssetsOnUninitialize { get; }