+++ Packages/UGF.Module.Assets/Runtime/IAssetTracker.cs
@@ -7,7 +7,7 @@ namespace UGF.Module.Assets.Runtime
         AssetTrack Add(string id, object asset);
         void Update(string id, AssetTrack track);
         bool Track(string id, object asset, out AssetTrack track);
-        bool UnTrack(string id, out AssetTrack track);
+        bool UnTrack(string id, object asset, out AssetTrack track);
         AssetTrack Increment(string id);
         AssetTrack Decrement(string id);
         AssetTrack GetByAsset(object asset);