+++ Packages/UGF.Module.Assets/Runtime/Loaders.Resources/ResourcesAssetGroupAsset.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using UGF.EditorTools.Runtime.IMGUI.Attributes;
 using UnityEngine;
 
 namespace UGF.Module.Assets.Runtime.Loaders.Resources
@@ -7,8 +8,11 @@ namespace UGF.Module.Assets.Runtime.Loaders.Resources
     [CreateAssetMenu(menuName = "Unity Game Framework/Assets/Resources Asset Group", order = 2000)]
     public class ResourcesAssetGroupAsset : AssetGroupAsset
     {
+        [AssetGuid(typeof(AssetLoaderAsset))]
+        [SerializeField] private string m_loader;
         [SerializeField] private List<Entry> m_assets = new List<Entry>();
 
+        public string Loader { get { return m_loader; } set { m_loader = value; } }
         public List<Entry> Assets { get { return m_assets; } }
 
         [Serializable]
@@ -21,14 +25,14 @@ namespace UGF.Module.Assets.Runtime.Loaders.Resources
             public string Address { get { return m_address; } set { m_address = value; } }
         }
 
-        protected override void OnPopulateGroup(IAssetGroup group)
+        protected override void OnGetAssets(IDictionary<string, IAssetInfo> assets)
         {
             for (int i = 0; i < m_assets.Count; i++)
             {
                 Entry entry = m_assets[i];
-                var info = new AssetInfo(entry.Address);
+                var info = new AssetInfo(m_loader, entry.Address);
 
-                group.Add(entry.Id, info);
+                assets.Add(entry.Id, info);
             }
         }
     }